Add params, load and save methods to PhysicalParams protocol.
Implement this for FullPhysicalParams.

We need to be able to save and load simulation parameters.
The params method should return a dictionary of the knwargs needed
to initialise a PhysicalParams object. The save method should
store these kwargs in a human readable format in the data/ directory.
The load classmethod should be able to instantiate a PhysicalParams
object from this stored file. These methods are implemented for
FullPhysicalParams using json file format.