Change definition of buoyancy parameter to use pore throat scale.
This means it is the same as the paper draft and does not depend on the
bubble size. This requires changing the definition of gas darcy
velocity.

Also fixed scaled bubble radius should depend on porosity not liquid
fraction.

Removed the coupling to the liquid flow in the calculation of gas darcy
velocity.

Removed static settings for DRAG_EXPONENT and PORE_THROAT_SCALING. The
pore throat scaling is set in the params and non dimensional params
objects now.